+++ src/ResourceManager/Compute/Commands.Compute.Test/Common/ComputeTestController.cs
@@ -14,7 +14,6 @@
 
 using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Management.Compute;
-using Microsoft.Azure.Management.Storage;
 using Microsoft.Azure.Test.HttpRecorder;
 using Microsoft.WindowsAzure.Commands.ScenarioTest;
 using Microsoft.WindowsAzure.Commands.Test.Utilities.Common;
@@ -25,9 +24,7 @@ using System.IO;
 using System.Linq;
 using Microsoft.Azure.Management.Network;
 #if NETSTANDARD
-using Microsoft.Azure.Management.ResourceManager;
 #else
-using Microsoft.Azure.Management.Resources;
 using Microsoft.Azure.Test;
 using TestBase = Microsoft.Azure.Test.TestBase;
 #endif
@@ -36,6 +33,9 @@ using Microsoft.Rest.ClientRuntime.Azure.TestFramework;
 using NetworkManagementClientInternal = Microsoft.Azure.Management.Internal.Network.Version2017_10_01.NetworkManagementClient;
 using ResourceManagementClientInternal = Microsoft.Azure.Management.Internal.Resources.ResourceManagementClient;
 using TestEnvironmentFactory = Microsoft.Rest.ClientRuntime.Azure.TestFramework.TestEnvironmentFactory;
+using Microsoft.Azure.Management.Internal.Resources;
+using CommonStorage = Microsoft.Azure.Management.Storage.Version2017_10_01;
+using PublicStorage = Microsoft.Azure.Management.Storage;
 
 namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
 {
@@ -43,7 +43,9 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
     {
         private readonly EnvironmentSetupHelper _helper;
 
-        public StorageManagementClient StorageClient { get; private set; }
+        public CommonStorage.StorageManagementClient StorageClient { get; private set; }
+
+        public PublicStorage.StorageManagementClient PublicStorageClient { get; private set; }
 
         public ComputeManagementClient ComputeManagementClient { get; private set; }
 
@@ -164,6 +166,7 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
             NetworkManagementClientInternal = GetNetworkManagementClientInternal(context);
             ResourceManagementClient = GetResourceManagementClient(context);
             InternalResourceManagementClient = GetResourceManagementClientInternal(context);
+            PublicStorageClient = GetPublicStorageManagementClient(context); 
 
             _helper.SetupSomeOfManagementClients(
                 StorageClient,
@@ -171,16 +174,13 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
                 NetworkManagementClient,
                 NetworkManagementClientInternal,
                 ResourceManagementClient,
-                InternalResourceManagementClient);
+                InternalResourceManagementClient,
+                PublicStorageClient);
         }
 
         private static ResourceManagementClient GetResourceManagementClient(MockContext context)
         {
-#if NETSTANDARD
             return context.GetServiceClient<ResourceManagementClient>(TestEnvironmentFactory.GetTestEnvironment());
-#else
-            return TestBase.GetServiceClient<ResourceManagementClient>(new CSMTestEnvironmentFactory());
-#endif
         }
 
         private static ResourceManagementClientInternal GetResourceManagementClientInternal(MockContext context)
@@ -188,9 +188,14 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
             return context.GetServiceClient<ResourceManagementClientInternal>(TestEnvironmentFactory.GetTestEnvironment());
         }
 
-        private static StorageManagementClient GetStorageManagementClient(MockContext context)
+        private static CommonStorage.StorageManagementClient GetStorageManagementClient(MockContext context)
         {
-            return context.GetServiceClient<StorageManagementClient>(TestEnvironmentFactory.GetTestEnvironment());
+            return context.GetServiceClient<CommonStorage.StorageManagementClient>(TestEnvironmentFactory.GetTestEnvironment());
+        }
+
+        private static PublicStorage.StorageManagementClient GetPublicStorageManagementClient(MockContext context)
+        {
+            return context.GetServiceClient<PublicStorage.StorageManagementClient>(TestEnvironmentFactory.GetTestEnvironment());
         }
 
         private static NetworkManagementClient GetNetworkManagementClient(MockContext context)