+++ src/ResourceManager/Compute/Commands.Compute/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Move dependencies on Keyvault and Storage to the common dependencies.
 * Add examples for `Remove-AzureRmVmssExtension`
 * Update help for `Set-AzureRmVMAccessExtension`